Silenced some warnings

+++ include/Cubism/SpaceFillingCurve.h
@@ -199,7 +199,7 @@ public:
     if (base_level < (double) (log(n_max) / log(2)) ) base_level ++;
     
     std::vector < std::array<unsigned int,3> > Indices(BX*BY*BZ);  
-    for (int h = 0 ; h < BX*BY*BZ ; h ++)
+    for (size_t h = 0 ; h < BX*BY*BZ ; h ++)
       TransposetoAxes(h,&Indices[h][0],base_level);
 
     for (unsigned int k=0;k<BZ;k++)
@@ -208,10 +208,10 @@ public:
     {
       const unsigned int c[3] = {i,j,k};
       
-      int index = AxestoTranspose( c, base_level);
+      size_t index = AxestoTranspose( c, base_level);
 
       int substract = 0;
-      for (int h=0; h<index; h++)
+      for (size_t h=0; h<index; h++)
       {
         if (Indices[h][0] >= BX ||  
             Indices[h][1] >= BY ||
